Fix frontend TypeScript issues and remove unused imports/logs
