+++ frontend/src/pages/Remote.tsx
@@ -16,8 +16,7 @@ const Remote = () => {
 	const navigate = useNavigate();
 	const { user } = useAuth();
 	const { friends, loading: friendsLoading } = useFriendsWithStatus(user?.id);
-	    //  console.log('🟢 FRIENDS WITH STATUS:', friends);
-    const [selectedFriend, setSelectedFriend] = useState<Friend | null>(null);
+	const [selectedFriend, setSelectedFriend] = useState<Friend | null>(null);
 	const [incomingInvite, setIncomingInvite] = useState<InvitePopupState>(null);
 	const { send, lastRawMessage, isConnected } = useNotificationSocket(true);
 
@@ -93,7 +92,6 @@ const Remote = () => {
 				</p>
 			) : (
 				<div className='w-full max-w-xl space-y-4'>
-
 					{friends.map((friend) => (
 						<div
 							key={friend.id}
