+++ services/game-service/.env.example
@@ -6,7 +6,7 @@ USER_SERVICE_URL=http://user:3003/api/users/state
 
 # Redis connection
 REDIS_HOST=redis
-REDIS_PORT=6379 #must be 79, this is inside the container
+REDIS_PORT=6379 # default Redis port inside the container
 
 # Presence service URL
 PRESENCE_SERVICE_URL=http://presence:3005
