+++ services/game-service/src/redis/subscribers.js
@@ -1,14 +1,13 @@
 import Redis from "ioredis";
-import { WebSocket } from "ws";
 
 const redisSubscriber = new Redis({
-  //Check if all these envs can be replaced by process.envREDIS_URL
   host: process.env.REDIS_HOST || "redis",
   port: process.env.REDIS_PORT ? parseInt(process.env.REDIS_PORT) : 6379,
   password: process.env.REDIS_PASSWORD || undefined,
 });
+
 // Subscribe to presence updates from Redis
-redisSubscriber.subscribe("presence:updates", (err: any, count: any) => {
+redisSubscriber.subscribe("presence:updates", (err, count) => {
   if (err) {
     console.error("[GAME-SERVICE] Failed to subscribe to presence:updates:", err);
   } else {
@@ -16,13 +15,11 @@ redisSubscriber.subscribe("presence:updates", (err: any, count: any) => {
   }
 });
 
-redisSubscriber.on("message", (channel: string, message: string) => {
+redisSubscriber.on("message", (channel, message) => {
   if (channel === "presence:updates") {
     try {
       const data = JSON.parse(message);
       console.log(`[GAME-SERVICE] Received presence update: ${data.userId} -> ${data.state}`);
-      
-      
     } catch (err) {
       console.error("[GAME-SERVICE] Redis message parse error:", err);
     }
@@ -32,5 +29,5 @@ redisSubscriber.on("message", (channel: string, message: string) => {
 export const redisPublisher = new Redis({
   host: process.env.REDIS_HOST || "redis",
   port: process.env.REDIS_PORT ? parseInt(process.env.REDIS_PORT) : 6379,
-password: process.env.REDIS_PASSWORD || undefined,
+  password: process.env.REDIS_PASSWORD || undefined,
 });
